Se cambia altoPieza de short a uint8_t. Se usa operador ternario para restar líneas de código

+++ mostrarPieza.h
@@ -3,6 +3,6 @@
 
 #include <cstdint>
 
-void mostrarPieza(uint8_t* pieza, short altoPieza);
+void mostrarPieza(uint8_t* pieza, uint8_t altoPieza);
 
 #endif
